pyproject.toml
[tool.poetry]
name = "card-generation-service"
version = "0.1.0"
description = ""
authors = ["Moritz Eich <hey@moritz.dev>"]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.104.1"
openai = "^0.27.0"
python-dotenv = "^0.19.2"
slowapi = "^0.1.5"
motor = "^3.1.2"
requests = "^2.27.1"
beautifulsoup4 = "^4.11.2"
readability-lxml = "^0.8.1"
transformers = "^4.27.4"
nltk = "^3.8.1"
tiktoken = "^0.4.0"
torch = [{version="^2.0.0", markers = "platform_machine == 'arm64'"},
{url="https://mirror.sjtu.edu.cn/pytorch-wheels/cpu-cxx11-abi/torch-2.0.0+cpu.cxx11.abi-cp311-cp311-linux_x86_64.whl", markers = "platform_machine == 'x86_64'"}]
httpx = "^0.24.1"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
lxml = "^4.9.3"
chromadb = "^0.4.18"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
boto3 = "^1.33.9"
pypandoc = "^1.12"
pandoc = "^2.3"
pydantic = "1.10.7"
modal = "0.62.21"
pymupdf = "^1.24.0"
pypdf = "^4.1.0"


[tool.poetry.group.dev.dependencies]
motor-stubs = "^1.7.1"
pip = "23.1.2"
black = "^22.1.0"
mypy = "^0.950"
types-requests = "^2.27.25"
pytest = "^7.1.2"
mongo-types = "^0.15.1"
install = "^1.3.5"
types-boto3 = "^1.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
check_untyped_defs = true